requirements.txt
environs~=9.5.0
python-telegram-bot~=13.13
google-cloud-dialogflow~=2.15.1
